requirements.txt
pandas
sqlalchemy
psycopg2
requests
plotly
dash
python-dotenv